requirements.txt
asgiref==3.3.1
astroid==2.5.1
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
docopt==0.6.2
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.16.0
pytz==2021.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
watchdog==2.0.2
whitenoise==5.2.0
wrapt==1.12.1
